+++ tests/testthat/test_AutoTuner.R
@@ -564,16 +564,48 @@ test_that("AutoTuner errors when second test set is not a subset of task ids", {
 })
 
 test_that("marshalable learner", {
-  task = tsk("mtcars")
+  task = tsk("iris")
   at = auto_tuner(
     tuner = tnr("random_search", batch_size = 2),
-    learner = lrn("regr.debug"),
+    learner = lrn("classif.debug"),
     resampling = rsmp("cv", folds = 3),
     measure = msr("classif.ce"),
-    term_evals = 4
+    term_evals = 4,
+    store_tuning_instance = TRUE
+  )
+  expect_true("marshal" %in% at$properties)
+
+  at$train(task)
+  at$marshal()
+  at$unmarshal()
+  expect_false(at$learner$marshaled)
+
+  expect_learner(at, task = task)
+})
+
+test_that("marshal", {
+  task = tsk("iris")
+  at = auto_tuner(
+    tuner = tnr("random_search", batch_size = 2),
+    learner = lrn("classif.debug"),
+    resampling = rsmp("cv", folds = 3),
+    measure = msr("classif.ce"),
+    term_evals = 4,
+    store_tuning_instance = TRUE
   )
 
   at$train(task)
 
-  expect_marshalable_learner(at, task)
+  model = at$model
+  model1 = marshal_model(model)
+  model2 = unmarshal_model(model1)
+  expect_false(model$learner$marshaled)
+  expect_true(model1$marshaled$learner$marshaled)
+  expect_false(model2$learner$marshaled)
+  expect_equal(class(model), class(model2))
+  expect_class(model1, "marshaled")
+
+  expect_false(identical(model$tuning_instance, model2$tuning_instance))
+  expect_false(identical(model$tuning_instance, model1$marshaled$tuning_instance))
+  expect_false(identical(model2$tuning_instance, model1$marshaled$tuning_instance))
 })